sitemap_no_search_version/requirements.txt
psycopg2-binary
python-dotenv
scrapy
nodriver
random_user_agent
geoip2
jsonschema